@@ -60,7 +60,9 @@ class AppStoreConnection implements InAppPurchaseConnection {
         productIdentifier: purchaseParam.productDetails.id,
         quantity: 1,
         applicationUsername: purchaseParam.applicationUserName,
-        simulatesAskToBuyInSandbox: purchaseParam.simulatesAskToBuyInSandbox,
+        // ignore: deprecated_member_use_from_same_package
+        simulatesAskToBuyInSandbox: purchaseParam.simulatesAskToBuyInSandbox ||
+            purchaseParam.sandboxTesting,
         requestData: null));
     return true; // There's no error feedback from iOS here to return.
   }
